[#121]refactoring: 클래스명 변경 및 분리
